+++ README.md
@@ -59,8 +59,22 @@ TransferFile(path, file.Item1, file.Item3);
 
 # Reference
 
+## Address
+*Values for the Address fields*
+
+Enum    | Example   | Description
+---     | ---:      | ---
+Cell    | A1        | Cell relative address
+CellAbs | A1        | Cell absolute address
+Range   | A1:B1     | Range relative address
+RangeAbs| A1:B1     | Range absoulute address
+RowOrCol| A or 1    | Row or Column address
+ColToCol| A:B       | Column to Column address
+RowToRow| 1:2       | Row to Row addess
+
+
 ## Commands
-*List of commands available*
+*List of available commands*
 
 Command     | Description | See Also
 ---         | --- | ---:
@@ -148,8 +162,8 @@ w   | false             | Wrap the text
 ## CellValueKind
 *Values for the CellValue command*
 
-Enum | Description
---- | ---
+Enum            | Description
+---             | ---
 Value           | Set the range to a specific value
 Text            | Returns the formatted value.
 AutoFilter      | Set an autofilter for the range
@@ -170,20 +184,20 @@ StyleName       | The named style
 ## ColumnValueKind
 *Values for the ColumnValue command*
 
-Enum | Description
---- | ---
+Enum            | Description
+---             | ---
 AutoFit         | Set the column width from the content of the range. The minimum width is the value of the ExcelWorksheet.defaultColumnWidth property. Note: Cells containing formulas are ignored since EPPlus don't have a calculation engine. Wrapped and merged cells are also ignored.
 BestFit         | If set to true a column automaticlly resize(grow wider) when a user inputs numbers in a cell.
 Merged          | none
 Width           | Sets the width of the column in the worksheet
-TrueWidth^      | Set width to a scaled-value that should result in the nearest possible value to the true desired setting.
+TrueWidth       | Set width to a scaled-value that should result in the nearest possible value to the true desired setting.
 
 
 ## ConditionalFormattingKind
 *Values for the ConditionalFormatting command*
 
-Enum | Description
---- | ---
+Enum            | Description
+---             | ---
 AboveAverage    | Add AboveAverage Rule
 AboveOrEqualAverage | Add AboveOrEqualAverage Rule
 AboveStdDev     | Add AboveStdDev Rule
@@ -234,8 +248,8 @@ Yesterday       | Add Yesterday Rule
 ## RowValueKind
 *Values for the RowValue command*
 
-Enum | Description
---- | ---
+Enum            | Description
+---             | ---
 Value           | Set the range to a specific value
 Text            | Returns the formatted value.
 AutoFilter      | Set an autofilter for the range
@@ -256,8 +270,8 @@ StyleName       | The named style
 ## ViewActionKind
 *Values for the ViewAction command*
 
-Enum | Description
---- | ---
+Enum            | Description
+---             | ---
 FreezePane      | Freeze the columns/rows to left and above the cell
 SetTabSelected  | Sets whether the worksheet is selected within the workbook.
 UnfreezePane    | Unlock all rows and columns to scroll freely
